# crates/dagron-engine/openapi.yaml
openapi: 3.0.3
info:
  title: module-54 workflow scheduler — management API
  description: >
    Read/write control surface over the workflow scheduler's datastore: list and
    inspect runs, submit a DAG, cancel a run, and scrape metrics. Every operation
    is a thin shell over the same `db` facade the reconcile loop uses, so the API
    holds no scheduler state of its own and works whether it runs in-process next
    to the loop or as a standalone read sidecar against the shared Postgres
    backend.


    Served only when the binary is built with the (default) `ops` Cargo feature
    and `API_ADDR` is set. This document is also served live by the running
    scheduler at `/openapi.yaml` and `/openapi.json`, with a Swagger UI at `/docs`.


    Security: this embedded API performs **no authentication** of its own — bind
    `API_ADDR` to localhost / a trusted network and place authentication (e.g. the
    `dagron-api` service, an auth proxy, or network policy) in front of it before
    exposing it. It is intended for in-process ops and trusted-network sidecar use.
  version: 0.1.0
  license:
    name: lab crate
servers:
  - url: http://127.0.0.1:8080
    description: Local management API (set API_ADDR to choose host:port)
...
paths:
...
  /runs/{id}/tasks/{task_id}/logs:
    get:
      tags: [runs]
      summary: Tail a task's output
      description: >
        One task's output for live tailing (fast-win #17). With `?offset=`
        returns only the output past that character offset, plus `next_offset`
        (resume point) and `eof` (the task is terminal); poll with
        `?offset=next_offset` until `eof`. Offsets are Unicode-scalar counts, so
        they never split a multibyte character.
      operationId: taskLogs
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/TaskId"
        - name: offset
          in: query
          required: false
          description: Resume tailing from this character offset. Omit for the full output.
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: The task's output (or the slice past `offset`) with tail metadata
          content:
            application/json:
              schema:
                type: object
                required: [task_id, name, status, attempt, output, offset, next_offset, eof]
                properties:
                  task_id: { type: string, format: uuid }
                  name: { type: string }
                  status: { $ref: "#/components/schemas/TaskStatus" }
                  attempt: { type: integer, format: int64 }
                  output: { type: string }
                  offset: { type: integer, description: char offset this slice starts at }
                  next_offset: { type: integer, description: resume point for the next poll }
                  eof: { type: boolean, description: true once the task is terminal }
        "404":
          description: Run or task not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          $ref: "#/components/responses/Error"
...
  /datasets:
    get:
      tags: [datasets]
      summary: List datasets
      description: >
        The dataset registry (data-aware scheduling, docs/DATASETS.md): every
        dataset a `produces:` task (or, with dagron Enterprise, an external
        event) ever updated, with its latest update, most recent first.
      operationId: listDatasets
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, format: int64, default: 100, minimum: 1, maximum: 1000 }
      responses:
        "200":
          description: Registered datasets
          content:
            application/json:
              schema:
                type: object
                required: [datasets]
                properties:
                  datasets:
                    type: array
                    maxItems: 1000
                    items: { $ref: "#/components/schemas/Dataset" }
        "500":
          $ref: "#/components/responses/Error"
  /datasets/events:
    get:
      tags: [datasets]
      summary: List dataset update events (lineage)
      description: >
        The append-only lineage ledger: which workflow/run/task updated which
        dataset when — the cross-workflow update trail dataset sensors and
        `on_datasets:` triggers key off. Newest first; `?uri=` narrows to one
        dataset.
      operationId: listDatasetEvents
      parameters:
        - name: uri
          in: query
          required: false
          schema: { type: string, maxLength: 512 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, format: int64, default: 100, minimum: 1, maximum: 1000 }
      responses:
        "200":
          description: Dataset update events
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    maxItems: 1000
                    items: { $ref: "#/components/schemas/DatasetEvent" }
        "500":
          $ref: "#/components/responses/Error"
    post:
      tags: [datasets]
      summary: Record an external dataset event (Enterprise)
      description: >
        Declare that a dataset was updated by a producer outside dagron (CDC,
        an object-store notification, another orchestrator), waking dataset
        sensors and firing `on_datasets:` triggers. Ships with dagron
        Enterprise; the open build answers `403` with a signpost — its datasets
        update via `produces:` tasks (docs/DATASETS.md).
      operationId: postDatasetEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [uri]
              properties:
                uri: { type: string, maxLength: 512, example: "s3://lake/orders" }
      responses:
        "200":
          description: Event recorded
          content:
            application/json:
              schema:
                type: object
                required: [recorded]
                properties:
                  recorded: { type: string }
        "400":
          description: Invalid dataset URI
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: External dataset events require dagron Enterprise (signposted)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          $ref: "#/components/responses/Error"
components:
  parameters:
    RunId:
      name: id
      in: path
      required: true
      description: Run UUID.
      schema: { type: string, format: uuid }
    TaskId:
      name: task_id
      in: path
      required: true
      description: Task UUID within the run.
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: Unexpected server/datastore error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Human-readable failure message }
...
    Dataset:
      type: object
      description: A registry row — one dataset URI with its latest update.
      required: [uri, updated_at, updates]
      properties:
        uri: { type: string, maxLength: 512, example: "s3://lake/orders" }
        updated_at: { type: string, format: date-time }
        last_run_id: { type: string, nullable: true, description: Run whose task last updated it (null for external events) }
        last_task: { type: string, nullable: true, description: Task name that last updated it }
        updates: { type: integer, format: int64, description: Total updates ever recorded }
    DatasetEvent:
      type: object
      description: One lineage-ledger row — a single recorded dataset update.
      required: [id, uri, source, at]
      properties:
        id: { type: integer, format: int64, description: Monotonic ledger id (sensor/trigger cursor domain) }
        uri: { type: string, maxLength: 512 }
        workflow: { type: string, nullable: true, description: Producing workflow (null for external events) }
        run_id: { type: string, nullable: true }
        task_name: { type: string, nullable: true }
        source: { type: string, enum: [task, external], description: "`task` = a produces: success; `external` = posted via the API (Enterprise)" }
        at: { type: string, format: date-time }
...
    TaskStatus:
      type: string
      description: Task state-machine node.
      enum: [pending, ready, running, awaiting_approval, succeeded, failed, skipped, cancelled]
...